APT_parser: Support custom APT file where the name of the internal .xml file is different from the name of the .aptx file
In the APT, the name of the .xml file stored in the .aptx file is supposed to have the same basename. But some people can save .aptx files locally, then rename it.

For instance:
* Save .aptx locally (e.g 1232.aptx)
* Rename it to *1232_new.aptx*

In that case, the name of the internal .xml file is still *1232.xml* and NOT *1232_new.xml*.

This commit provide a way to retrieve the .xml file from the file list of the .aptx file, assuming it is the only .xml file in the list (which is the case at the moment with APT v27.2)