Poisson Noise Changes
These changes rework the way Poisson noise is handled. In each interval of time the number of events is independent of the current number of counts. Thus, the change in the accumlated counts is just a Poisson sample of the expected counts in the interval.  I had to make some other changes to not lose track of the cosmic ray events.